user prompt: update product price at store (fixed getting price)

+++ app/src/main/java/com/example/shopist/Server/ServerResponses/ServerShoppingProduct.java
@@ -5,14 +5,12 @@ public class ServerShoppingProduct extends ServerProduct {
     //this product has no price, since it is not attached to a store(? may be ?)
 
     private int needed;
-    private int price;
+    private float price;
 
     public int getNeeded() {
         return needed;
     }
 
-    public int getPrice() {
-        return price;
-    }
+    public float getPrice() { return price; }
 
 }
